Removed dead code with devnet-2 feature flags, removed unused imports in validator/lib.rs and moved the unit_tests to separate folder and file
