+++ lean_client/validator/src/lib.rs
@@ -9,7 +9,7 @@ use containers::{
     block::{hash_tree_root, BlockWithAttestation, SignedBlockWithAttestation},
     checkpoint::Checkpoint,
     types::{Uint64, ValidatorIndex},
-    AggregatedAttestation, Slot,
+    Slot,
 };
 use fork_choice::store::{get_proposal_head, get_vote_target, Store};
 use tracing::{info, warn};
@@ -191,12 +191,6 @@ impl ValidatorService {
             .map(|att| att.message.clone())
             .collect();
 
-        #[cfg(feature = "devnet2")]
-        let valid_attestations: Vec<AttestationData> = valid_signed_attestations
-            .iter()
-            .map(|att| att.message.clone())
-            .collect();
-
         info!(
             slot = slot.0,
             valid_attestations = valid_attestations.len(),
@@ -325,14 +319,6 @@ impl ValidatorService {
                     source: store.latest_justified.clone(),
                 };
 
-                #[cfg(feature = "devnet2")]
-                let attestation = AttestationData {
-                    slot,
-                    head: head_checkpoint.clone(),
-                    target: vote_target.clone(),
-                    source: store.latest_justified.clone(),
-                };
-
                 let signature = if let Some(ref key_manager) = self.key_manager {
                     // Sign with XMSS
                     let message = hash_tree_root(&attestation);
@@ -377,47 +363,5 @@ impl ValidatorService {
                 })
             })
             .collect()
-    }
-}
-
-// DI GENERUOTI TESTAI
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_proposer_selection() {
-        let config = ValidatorConfig {
-            node_id: "test_0".to_string(),
-            validator_indices: vec![2],
-        };
-        let service = ValidatorService::new(config, 4);
-
-        // Validator 2 should propose at slots 2, 6, 10, ...
-        assert!(service.get_proposer_for_slot(Slot(2)).is_some());
-        assert!(service.get_proposer_for_slot(Slot(6)).is_some());
-        assert!(service.get_proposer_for_slot(Slot(10)).is_some());
-
-        // Validator 2 should NOT propose at slots 0, 1, 3, 4, 5, ...
-        assert!(service.get_proposer_for_slot(Slot(0)).is_none());
-        assert!(service.get_proposer_for_slot(Slot(1)).is_none());
-        assert!(service.get_proposer_for_slot(Slot(3)).is_none());
-        assert!(service.get_proposer_for_slot(Slot(4)).is_none());
-        assert!(service.get_proposer_for_slot(Slot(5)).is_none());
-    }
-
-    #[test]
-    fn test_is_assigned() {
-        let config = ValidatorConfig {
-            node_id: "test_0".to_string(),
-            validator_indices: vec![2, 5, 8],
-        };
-
-        assert!(config.is_assigned(2));
-        assert!(config.is_assigned(5));
-        assert!(config.is_assigned(8));
-        assert!(!config.is_assigned(0));
-        assert!(!config.is_assigned(1));
-        assert!(!config.is_assigned(3));
     }
 }
